+++ package.json
@@ -12,13 +12,15 @@
   "dependencies": {
     "@tanstack/react-query": "^5.99.1",
     "firebase": "^12.12.0",
+    "jotai": "^2.19.1",
     "jspdf": "^4.2.1",
     "react": "^19.2.4",
     "react-day-picker": "^9.14.0",
     "react-dom": "^19.2.4",
     "react-icons": "^5.6.0",
     "react-router-dom": "^7.14.1",
-    "react-spinners": "^0.17.0"
+    "react-spinners": "^0.17.0",
+    "sass": "^1.99.0"
   },
   "devDependencies": {
     "@babel/core": "^7.29.0",
