+++ src/App.scss
@@ -1,0 +1,33 @@
+@use "./variables" as *;
+@import url("https://fonts.googleapis.com/css?family=Lobster+Two");
+
+:root {
+  --text: #{$fg-main};
+  --background: #{$bg-main};
+  --backgroundTwo: #{$bg-secondary};
+  --backgroundThree: #{$bg-third};
+  --primary: #{$primary};
+  --secondary: #{$secondary};
+  --accent: #{$accent};
+
+  --titleFont: #{$title-font};
+
+  font-size: 16px;
+}
+
+// set up some defaults
+* {
+  box-sizing: border-box;
+}
+
+button {
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  border: none;
+  border-radius: 3px;
+}
+
+.loading-page {
+  width: 100%;
+  height: 100%;
+  background-color: green;
+}
