+++ src/pages/auth/Login.css
@@ -45,6 +45,7 @@
   margin-bottom: 10px;
   border: 1px solid #ccc;
   border-radius: 3px;
+  box-sizing: border-box;
 }
 
 .login-button {
