+++ src/pages/auth/sass/Login.module.scss
@@ -1,0 +1,8 @@
+@use "../../../variables" as *;
+
+.login-button {
+  @include primary-button;
+
+  width: 100%;
+  padding: 0.75rem;
+}
