+++ src/pages/dashboard/OmniAdmin.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
-import { performLogout, transferEmpData } from "../utils/firebase.ts";
+import { performLogout, transferEmpData } from "../../utils/firebase.ts";
 import "react-day-picker/dist/style.css";
 import "./OmniAdmin.css";
-import DisplayTable from "../components/DisplayTable.tsx";
-import ClickBlocker from "../components/ClickBlocker.tsx";
-import { useCompanies } from "../utils/firebaseQueries.ts";
+import DisplayTable from "../../components/DisplayTable.tsx";
+import ClickBlocker from "../../components/ClickBlocker.tsx";
+import { useCompanies } from "../../utils/firebaseQueries.ts";
 
 function OmniAdminDashboard(props) {
   const [oldUserID, setOldUserID] = useState("");
