+++ src/utils/atoms.tsx
@@ -1,0 +1,4 @@
+import { atom } from "jotai";
+
+export const UID = atom("");
+export const ResetToken = atom("");
