+++ src/variables.scss
@@ -1,0 +1,24 @@
+$fg-main: #0e1b08;
+
+$bg-main: #f7fcf5;
+$bg-secondary: #f7f7f7;
+$bg-third: #f1f1f1;
+
+$primary: #54bc55;
+$secondary: #9dd7d9;
+$accent: #76a8c9;
+
+$title-font: "Lobster Two";
+
+@mixin primary-button {
+  background-color: $primary;
+  color: $bg-main;
+  border: none;
+  border-radius: 0.2rem;
+  cursor: pointer;
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+
+  &:hover {
+    background-color: $secondary;
+  }
+}
